pyproject.toml
[tool.poetry]
name = "bedrock-playground"
version = "0.1.0"
description = "Playing with Bedrock API"
authors = ["Noah Sleiman <nksleiman@gmail.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.34.23"
requests = "^2.31.0"
streamlit = "^1.30.0"
pyyaml = "^6.0.1"
langchain = "^0.1.1"
langchain-community = "^0.0.13"
transformers = "^4.36.2"
atlassian-python-api = "^3.41.9"
lxml = "^5.1.0"
pytesseract = "^0.3.10"
pdf2image = "^1.17.0"
flatbuffers = "^23.5.26"
onnxruntime = "^1.16.3"
chromadb = "^0.4.22"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
ipython = "^8.20.0"

[build-system]
requires = ["poetry-core>=1.7.1"]
build-backend = "poetry.core.masonry.api"
